requirements.txt
fastapi
uvicorn
openai
chromadb
langchain
langchain-community
langchain-openai
pypdf
python-dotenv
requests